moved Notifier to JoeScan.LogScanner.Shared
