+++ src/JoeScan.LogScanner.LogReview/Models/LogReviewer.cs
@@ -2,8 +2,10 @@
 using Caliburn.Micro;
 using JoeScan.LogScanner.Core.Models;
 using JoeScan.LogScanner.LogReview.Interfaces;
+using JoeScan.LogScanner.Shared.Notifier;
 using NLog;
 using System;
+using System.Linq;
 
 namespace JoeScan.LogScanner.LogReview.Models;
 
@@ -11,6 +13,7 @@ public class LogReviewer : PropertyChangedBase, ILogModelObservable
 {
     public ILogger Logger { get; }
     public LogModelBuilder ModelBuilder { get; }
+    public IUserNotifier Notifier { get; }
 
     #region Bound Properties
 
@@ -27,7 +30,16 @@ public class LogReviewer : PropertyChangedBase, ILogModelObservable
             NotifyOfPropertyChange(() => CurrentRawLog);
             if (currentRawLog != null)
             {
-                CurrentLogModel = ModelBuilder.Build(currentRawLog).LogModel;
+                var res =  ModelBuilder.Build(currentRawLog);
+                if (res.IsValidModel)
+                {
+                    CurrentLogModel = res.LogModel;
+                }
+                else
+                {
+                    Notifier.Error($"Failed to build LogModel from RawLog. Error was: {String.Join(',',res.Messages)}");
+                    CurrentLogModel = null;
+                }
             }
         }
     }
@@ -74,10 +86,12 @@ public class LogReviewer : PropertyChangedBase, ILogModelObservable
     #region Lifecycle
 
     public LogReviewer(ILogger logger,
-        LogModelBuilder modelBuilder)
+        LogModelBuilder modelBuilder,
+        IUserNotifier notifier)
     {
         Logger = logger;
         ModelBuilder = modelBuilder;
+        Notifier = notifier;
     }
 
     #endregion
