+++ src/JoeScan.LogScanner.LogReview/nlog.config
@@ -3,32 +3,35 @@
 <nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
     <targets>
-        <target name="corelog" xsi:type="AsyncWrapper" queueLimit="5000" overflowAction="Discard">
-            <target xsi:type="File"
-                    fileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/LogScanner/logs/JoeScan.LogScanner.Core-${shortdate}.txt"
-                    archiveFileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/ThicknessMonitor/logs/archives/JoeScan.LogScanner.Core-${shortdate}.{#}.txt"
-                    archiveNumbering="Rolling"
-                    maxArchiveFiles="7"
-                    concurrentWrites="false"
-                    keepFileOpen="true"
-                    >
-            </target>
-        </target>
-        <target name="reviewlog" xsi:type="AsyncWrapper" queueLimit="5000" overflowAction="Discard">
-            <target xsi:type="File"
-                    fileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/LogScanner/logs/JoeScan.LogScanner.LogReview-${shortdate}.txt"
-                    archiveFileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/ThicknessMonitor/logs/archives/JoeScan.LogScanner.LogReview-${shortdate}.{#}.txt"
-                    archiveNumbering="Rolling"
-                    maxArchiveFiles="7"
-                    concurrentWrites="false"
-                    keepFileOpen="true"
-                   >
-            </target>
-        </target>
-    </targets>
+        <!-- <target name="corelog" xsi:type="AsyncWrapper" queueLimit="5000" overflowAction="Discard"> -->
+        <!--     <target xsi:type="File" -->
+        <!--             fileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/LogScanner/logs/JoeScan.LogScanner.Core-${shortdate}.txt" -->
+        <!--             archiveFileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/ThicknessMonitor/logs/archives/JoeScan.LogScanner.Core-${shortdate}.{#}.txt" -->
+        <!--             archiveNumbering="Rolling" -->
+        <!--             maxArchiveFiles="7" -->
+        <!--             concurrentWrites="false" -->
+        <!--             keepFileOpen="true" -->
+        <!--             > -->
+        <!--     </target> -->
+        <!-- </target> -->
+        <!-- <target name="reviewlog" xsi:type="AsyncWrapper" queueLimit="5000" overflowAction="Discard"> -->
+        <!--     <target xsi:type="File" -->
+        <!--             fileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/LogScanner/logs/JoeScan.LogScanner.LogReview-${shortdate}.txt" -->
+        <!--             archiveFileName="${specialfolder:folder=CommonApplicationData:cached=true}/JoeScan/ThicknessMonitor/logs/archives/JoeScan.LogScanner.LogReview-${shortdate}.{#}.txt" -->
+        <!--             archiveNumbering="Rolling" -->
+        <!--             maxArchiveFiles="7" -->
+        <!--             concurrentWrites="false" -->
+        <!--             keepFileOpen="true" -->
+        <!--            > -->
+        <!--     </target> -->
+        <!-- </target> -->
+	    <target name="console" xsi:type="Console" layout="${date:format=HH\:MM\:ss} ${logger} ${message}" />
+
+
+	</targets>
     <rules>
-        <logger name="JoeScan.LogScanner.Core.*" minlevel="Debug" writeTo="corelog" />
-        <logger name="JoeScan.LogScanner.LogReview.*" minlevel="Debug" writeTo="reviewlog" />
-        
+        <!-- <logger name="JoeScan.LogScanner.Core.*" minlevel="Debug" writeTo="corelog" /> -->
+        <!-- <logger name="JoeScan.LogScanner.LogReview.*" minlevel="Debug" writeTo="reviewlog" /> -->
+		<logger name="*" minlevel="Trace" writeTo="console" />
     </rules>
 </nlog>
