+++ src/JoeScan.LogScanner.Shared/Externals.xaml
@@ -1,0 +1,5 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary  Source="pack://application:,,,/ToastNotifications.Messages;component/Themes/Default.xaml" />
+    </ResourceDictionary.MergedDictionaries>
+</ResourceDictionary>
